M0 V2: extract the ramp curve and error classes, and label the axis honestly
V1 built the Load and Errors tab bodies inline. This extracts them into the three
components the node declares and fixes the one real defect it left behind.

The defect: the ramp chart's X axis said "concurrent clients" but the Y axis had
no label at all, so the headline curve rendered bare milliseconds an operator
could read as caller-visible answer latency. It is not. It is the prober's own
message through the SFU data channel and back. The Y axis now reads "SFU
data-channel RTT (ms)", the card title says the same, and the closing note spells
out in words that this is not the latency a caller hears. Answer latency arrives
later in M1/M2 with its own answer_latency_source provenance; RTT_AXIS_LABEL must
not be reused for it.

New: components/LoadCurveChart.tsx (chart only, no card wrapper, following the
LatencyWaterfall precedent), components/ErrorClassChart.tsx, and lib/errorClass.ts
holding the classifier and label/colour maps moved out of ErrorsTab. The tabs lose
202 lines and gain 26.

Behaviour is preserved, not re-derived: classifyKnee and KneeBanner are untouched,
so find_knee's two opposite None outcomes stay distinguished ("No knee inside this
ramp" vs "No healthy tier in this ramp"). The loss series stays suppressed and no
loss column is shown. No dependency added; recharts was already present.

Verified in the live DOM rather than only compiled: the Load tab's SVG text is
["2","10","25","concurrent clients","0","20","40","60","80",
"SFU data-channel RTT (ms)","50 ms budget","knee 10"], with no console errors.

Gates: frontend tsc+build+build:demo exit 0; behavioural MIN_CARDS=5 GREEN with 6
rendered cards. No Python touched, so no backend gate. No test file touched.
